# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaPhoto
  
  description: API documentation for the WasaPhoto project.
  version: 1.0.0
security:
  - BearerAuth: [] 
...
paths:
...
  /users/{username}/follows: 
    parameters:
    - $ref: "#/components/parameters/username"
    
    put:
      tags: ["User interaction"]
      
      operationId: followUser
      summary: follow a user
      description: add this user to the list of users our user follows
      responses:
        '200':
          description: User successfully followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '400':
          description: Incorrect request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '401':
          description: Action unauthorized - invalid user token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
                
    get:
     
      tags: ["User interaction"]
          
      operationId: getfollowers
      summary: retrive followers 
      description: get a list of followers for this user id. 
      
          
      responses: 
        "200": 
          description: list of followers
          content: 
            application/json: 
              schema: 
                type: array
                minItems: 0
                maxItems: 800
                items: { $ref: "#/components/schemas/user" }
        "401": 
          description: Unauthorized - Authentication failed or token expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"

        "500": 
          description: Internal Server Error - could be error of authentication 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "404": 
          description: user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
  
  
  #get following/ unfollow
  /users/{username}/follows/{target_username}:
    parameters:
    - $ref: "#/components/parameters/username"
    - $ref: "#/components/parameters/target_username"
    
    get:
      tags: ["User interaction"]
          
      operationId: getfollowing
      summary: retrive following
      description: get a list of users that this user follows. 
      
      responses: 
        "200": 
          description: list of following
          content: 
            application/json: 
              schema: 
                type: array
                minItems: 0
                maxItems: 800
                items: { $ref: "#/components/schemas/user" }  
        "401": 
          description: Unauthorized - Authentication failed or token expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "500": 
          description: Internal Server Error - could be error of authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "404": 
          description: user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
                
    
    delete:
      tags: ["User interaction"]
      
      operationId: unfollowUser
      summary: unfollow user
      description: Remove a user from the following list.
    
         
      responses:
      
        "204":  
          description: unfollowed 
        '400':
          description: Incorrect request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '401':
          description: Action unauthorized - invalid user token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
                    
   

  
 #ban/unban/getbanned 
  /users/{username}/bans/{target_username}:
    parameters:
    - $ref: "#/components/parameters/target_username"
    - $ref: "#/components/parameters/username"
    put:
      
      tags: ["User interaction"]
        
      operationId: banUser
      summary: Ban a user
      description: Add this user to the list of banned users for the current user
      
      responses:
        "204":
          description: User banned successfully
          
        "400":
          description: Bad request - Invalid input or target user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "403":
          description: Forbidden - You cannot ban yourself
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "500":
          description: Internal Server Error - Could be an issue with authentication or server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
    
    delete:
      tags: ["User interaction"]
      operationId: unbanUser
      summary: Unban a user
      description: Remove a user from the banned list.
      responses:
        '200':
          description: User successfully unbanned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '400':
          description: Incorrect request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '401':
          description: Action unauthorized - invalid user token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
                
                
                
    get: 
      operationId: getbannedusers
      summary: get banned users
      tags: ["User interaction"]
      description: retrive a list of banned users
      
      responses: 
          "200": 
            description: list retrived
            content:
              application/json:
                schema: 
                  type: array
                  minItems: 0
                  maxItems: 800
                  items: { $ref: "#/components/schemas/user" }

  
  
#get user profil
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: string
      description: "Use a Bearer token for authentication. Obtain your token from the authentication endpoint."
  schemas:
    user:
      title: "User"
      description: An object that represents a user of the application.
      type: object
      properties:
        username:
          description: "A username serving as a unique user ID"
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: "rodi"
...
    EmptyResponse:
      description: Returned when the respone is empty
      type: object 
  parameters:
    username:
      name: username
      in: path
      required: true
      description: "The name of the user"
      schema:
        description: "The name of the user"
        type: string
        pattern: '^.*?$'
        example: "mimi"
        minLength: 3
        maxLength: 16
    target_username:
      name: target_username
      in: path
      required: true
      description: "The name of the target user"
      schema:
        description: "The name of the target user"
        type: string
        pattern: '^.*?$'
        example: "mimi"
        minLength: 3
        maxLength: 16
...